object_detection/requirements.txt
torch
ultralytics
opencv-python
numpy
pyyaml